requirements.txt
Django==1.11.1
django-allauth==0.32.0
django-webpack-loader==0.5.0
django_debug_toolbar==1.8
django_mptt==0.8.7
djangorestframework==3.6.3
pandas==0.20.3  
psycopg2==2.7.1 
amqp==2.0.3  
backports-abc==0.4 
fastnumbers==0.7.3  
billiard==3.3.0.22
kombu==3.0.35 
numpy==1.10.4  
python-daemon==2.1.1
python-dateutil==2.5.0
pytz==2015.7 
raven==6.1.0
redis==2.10.5 
requests==2.9.1  
rpy2==2.7.8  
six==1.10.0
